agregado controles para que los Usuarios Simples no puedan interaccionar con nada qu eno sea de ellos

+++ Controllers/TableroController.cs
@@ -2,13 +2,14 @@ namespace Tp11.Controllers;
 
 using System.Diagnostics;
 using Microsoft.AspNetCore.Mvc;
+using System.Data.SQLite;
 
 using Tp11.Models;
 using Tp11.ViewModels;
 using EspacioTableroRepository;
 
 public class TableroController : Controller{
-    //TableroRepository repo = new TableroRepository();
+    private readonly string direccionBD = "Data Source = DataBase/kamban.db;Cache=Shared";
     private readonly ITableroRepository repo;
     private readonly ILogger<HomeController> _logger;
     public TableroController(ILogger<HomeController> logger, ITableroRepository TabRepo) //constructor de Tablero que recibe un parametro tipo ILogger<HomeController> 
@@ -26,7 +27,15 @@ public class TableroController : Controller{
             if (isAdmin()){
                 tableros = repo.GetAll();
             }else if(idUsuario.HasValue){
-                tableros = repo.GetTablerosDeUsuario(idUsuario);//ver como poner el parametro adecuado segun el id del usuario logueado
+                int? ID = ObtenerIDDelUsuarioLogueado(direccionBD);
+                if (ID == idUsuario)
+                {
+                    tableros = repo.GetTablerosDeUsuario(ID);
+                }else
+                {
+                    return NotFound();
+                }
+
             }else{
                 return NotFound();
             }
@@ -45,6 +54,7 @@ public class TableroController : Controller{
         try
         {
             if(!isLogin()) return RedirectToAction("Index","Login");
+            if(!isAdmin()) return RedirectToAction("Index","Login");
 
             TableroViewModel newTableroVM = new TableroViewModel();
             return View(newTableroVM);
@@ -61,10 +71,12 @@ public class TableroController : Controller{
         {
             if(!ModelState.IsValid) return RedirectToAction("Index","Login");
             if(!isLogin()) return RedirectToAction("Index","Login");
+            if(!isAdmin()) return RedirectToAction("Index","Login");
 
             Tablero newTablero = Tablero.FromCrearTableroViewModel(newTableroVM);
+            int? ID = newTablero.IdUsuarioPropietario;
             repo.Create(newTablero);
-            return RedirectToAction("Index");
+            return RedirectToAction("Index", new { idUsuario = ID });//redirecciona al index con el idDelUsuario en caso de que sea un usuario Simple
         }
         catch (Exception ex)
         {
@@ -76,11 +88,25 @@ public class TableroController : Controller{
     [HttpGet]
     public IActionResult EditarTablero(int? idTablero){
         try
-        {
+        {   
             if(!isLogin()) return RedirectToAction("Index","Login");
 
             Tablero tableroAEditar = repo.GetById(idTablero);
-            TableroViewModel tableroAEditarVM = TableroViewModel.FromTablero(tableroAEditar);
+            TableroViewModel tableroAEditarVM = null;
+            
+            if (isAdmin()){
+                 tableroAEditarVM = TableroViewModel.FromTablero(tableroAEditar);
+            }else if(idTablero.HasValue){
+                int? ID = ObtenerIDDelUsuarioLogueado(direccionBD);
+                
+                if (ID == tableroAEditar.IdUsuarioPropietario){
+                    tableroAEditarVM = TableroViewModel.FromTablero(tableroAEditar);
+                }else{
+                    return NotFound();
+                }
+            }else{
+                return NotFound();
+            }
             return View(tableroAEditarVM);
         }
         catch (Exception ex)
@@ -98,8 +124,9 @@ public class TableroController : Controller{
             if(!isLogin()) return RedirectToAction("Index","Login");
 
             Tablero tableroAEditar = Tablero.FromEditarTableroViewModel(tableroAEditarVM);
+            int? ID = tableroAEditar.IdUsuarioPropietario;
             repo.Update(tableroAEditar);
-            return RedirectToAction("Index");
+            return RedirectToAction("Index", new { idUsuario = ID });
         }
         catch (Exception ex)
         {
@@ -111,11 +138,26 @@ public class TableroController : Controller{
     [HttpGet]
     public IActionResult EliminarTablero(int? idTablero){
         try
-        {
+        {   
             if(!isLogin()) return RedirectToAction("Index","Login");
-
+            
             Tablero tableroAEliminar = repo.GetById(idTablero);
-            return View(tableroAEliminar);
+            int? idUsuarioTablero = tableroAEliminar.IdUsuarioPropietario;
+            
+            if (isAdmin()){
+                return View(tableroAEliminar);
+            }else if(idTablero.HasValue){
+                int? ID = ObtenerIDDelUsuarioLogueado(direccionBD);
+                
+                if (ID == idUsuarioTablero){
+                    return View(tableroAEliminar);
+                }else{
+                    return NotFound();
+                }
+            }else
+            {
+                return NotFound();
+            }
         }
         catch (Exception ex)
         {
@@ -125,13 +167,14 @@ public class TableroController : Controller{
         
     }
     [HttpPost]
-    public IActionResult EliminarFromForm(Tablero tableroAEliminar){
+    public IActionResult EliminarFromForm([FromForm] Tablero tableroAEliminar){
         try
         {
+            if(!ModelState.IsValid) return RedirectToAction("Index","Login");
             if(!isLogin()) return RedirectToAction("Index","Login");
-        
+            
             repo.Remove(tableroAEliminar.Id);
-            return RedirectToAction("Index");
+            return RedirectToAction("Index", "Usuario");
         }
         catch (Exception ex)
         {
@@ -156,6 +199,35 @@ public class TableroController : Controller{
         }else{
             return false;
         }
+    }
+
+    private int? ObtenerIDDelUsuarioLogueado(string? direccionBD){// se agrego para poder hacer el control cuando se seleccione editar/agregar/eliminar tableros que no son del usuario logueado y este no es Admin
+        int? ID = 0;
+        Usuario usuarioSelec = new Usuario();
+        SQLiteConnection connectionC = new SQLiteConnection(direccionBD);
+
+        string queryC = "SELECT * FROM Usuario WHERE nombre_de_usuario = @NAME AND contrasenia = @PASS";
+        SQLiteParameter parameterName = new SQLiteParameter("@NAME", HttpContext.Session.GetString("Nombre"));
+        SQLiteParameter parameterPass = new SQLiteParameter("@PASS", HttpContext.Session.GetString("Contrasenia"));
+
+        using (connectionC)
+        {
+            connectionC.Open();
+            SQLiteCommand commandC = new SQLiteCommand(queryC,connectionC);
+            commandC.Parameters.Add(parameterName);
+            commandC.Parameters.Add(parameterPass);
+            
+            SQLiteDataReader readerC = commandC.ExecuteReader();
+            using (readerC)
+            {
+                while (readerC.Read())
+                {
+                    ID = Convert.ToInt32(readerC["id"]);
+                }
+            }
+            connectionC.Close();
+        }
+        return(ID);
     }
 
     [ResponseCache(Duration = 0, Location = ResponseCacheLocation.None, NoStore = true)]
